+++ index.js
@@ -12,7 +12,6 @@ import { getKinstaHelpMessage } from "./help/kinsta.js"
 import { resetConf } from "./lib/conf.js"
 import { checkDependencies, checkNodeVersion } from "./lib/dependencies.js"
 import { getSitesPath } from "./lib/path.js"
-import setupAWS from "./tasks/services/aws.js"
 
 const program = new Command()
 
@@ -25,8 +24,6 @@ async function initProgram() {
   await checkDependencies()
 
   process.chdir(getSitesPath())
-
-  program.command("aws").action(() => setupAWS("production"))
 
   program
     .command("path")